+++ app/src/main/java/com/github/wanderwise_inc/app/data/GoogleSignInLauncher.kt
@@ -1,17 +1,5 @@
 package com.github.wanderwise_inc.app.data
 
-import android.content.Intent
-import androidx.activity.result.ActivityResultLauncher
-
 interface GoogleSignInLauncher {
   fun launchSignIn()
-}
-
-class DefaultGoogleSignInLauncher(
-    private val signInLauncher: ActivityResultLauncher<Intent>,
-    private val signInIntent: Intent
-) : GoogleSignInLauncher {
-  override fun launchSignIn() {
-    signInLauncher.launch(signInIntent)
-  }
 }
